Use metadata directory instead of index json file
Use the metadata directory, as the we always need the contents of the
metdata yml files. The previous code grabbed all metadata using https
via GitLab, which generates a lot of requests. Therefore it probably
already had a the feature to have a metadata repo checkout at
/tmp/fdroiddata to prevent the high request amount.

Now the archive checker is passed the directory of the F-Droid metadata
directory. This has the advantage to also check different versions of
the metadata repository or check other metadata repositories.