Check if repositories moved to Github or GitLab
Add a check for archived repositories, if they moved to Github or GitLab.
This uses the fact, that a moved repository should keep it's Git
history. This allows us to check, if the last F-Droid release commit is
contained in a potential new repository url. This check allows to simply
check all relevant URLs in the response, if they point to a new
repository.

Note that the result keeps the archived repository flag, but has an
additional field to indicate the new archive. After the metadata has
updated accordingly, it can be checked again (in case the new repository
is also archived).