Don't create false positives with Github forks
When a fork of a GitHub repository adds commmits, then the commits are
also accessible from the original repository. This is probably just due
to GitHub optimizing the required storage for forks. But Github shows
dynamically a warning, when the commit doesn't match a branch or tag in
the current repository:

    This commit does not belong to any branch on this repository, and
    may belong to a fork outside of the repository.

Therefore use the request Github uses to check this condition to check
if a commit is contained in a given repository url.